+++ languages/de_DE/messages.php
@@ -49,7 +49,7 @@
     "No Galleries found"=> "Keine Galerien gefunden",
     "{0} %count% Galleries|{1} %count% Gallery|]1,Inf[ %count% Galleries"=> "{0} %count% Galerien|{1} %count% Galerie|]1,Inf[ %count% Galerien",
     "{1} %count% Gallery selected|]1,Inf[ %count% Galleries selected" => "{1} %count% Galerie ausgewählt|]1,Inf[ %count% Galerien ausgewählt",
-    "{1} %count% File selected|]1,Inf[ %count% Files selected" => "{1} %count% File selected|]1,Inf[ %count% Files selected",
+    "{1} %count% File selected|]1,Inf[ %count% Files selected" => "{1} %count% Datei ausgewählt|]1,Inf[ %count% Dateien ausgewählt",
     "Invalid file type. Only *.jpg, *.jpeg and *.png are supported" => "Ungültiger Dateityp. Nur *.jpg, *.jpeg und *.png werden unterstützt",
     "General" => "Allgemein",
     "Images" => "Bilder",
@@ -83,5 +83,6 @@
     "Delete image" => "Bild löschen",
     "max %size%MB per upload" => "maximal %size%MB pro Upload",
     "more" => "mehr",
-    "Show Link?" => "Link anzeigen?"
+    "Show Link?" => "Link anzeigen?",
+    "Limit" => "Limit"
 );